refactor(geometry): split the ray-parity block out of classify.rs
classify.rs was 739 lines against its 720-line module-size ratchet
budget. Per louistrue's review on #3386, move the ray-parity block
(exact_seg_hits_tri, operand_extent, ray_dir, sound_far, point_inside)
into a new sibling module, arrangement/ray_parity.rs, rather than
raising the budget or trimming the soundness rationale in sound_far's
doc comment.

Pure move: no logic changed. ray_dir, exact_seg_hits_tri and sound_far
widen from module-private (classify.rs only) to pub(super) (visible
within the arrangement module), since classify.rs now calls them
across the new file boundary; visibility does not extend past
arrangement.

classify.rs: 739 -> 614 lines (budget 720).
ray_parity.rs: 141 lines (new, no allowlist row needed).

Not run here: the corpus census (1.4 GB of fixtures, doesn't fit this
environment). Must be run before merge, since this touches classify.rs.
