+++ rust/geometry/src/kernel/arrangement/mod.rs
@@ -31,6 +31,7 @@ use std::cmp::Ordering;
 
 mod boolean;
 mod classify;
+mod ray_parity;
 #[cfg(test)]
 mod tests;
 
@@ -38,7 +39,6 @@ pub use self::boolean::{
     boolean, boolean_manifest, boolean_topology_hash, box_mesh, cube_mesh, difference_all,
     difference_all_lenient, union_all,
 };
-
 pub type Tri = [[f64; 3]; 3];
 
 /// Boolean operation.
